+++ rust/datafusion-extensions/src/lib.rs
@@ -1,5 +1,7 @@
 /// Unified binary column accessor for Arrow arrays
 pub mod binary_column_accessor;
+/// Binning UDFs (bin_center)
+pub mod binning;
 /// Color UDFs (rgba, lerp_color)
 pub mod color;
 /// Compute histograms from SQL
@@ -11,6 +13,7 @@ pub mod properties;
 
 use std::sync::Arc;
 
+use binning::bin_center::make_bin_center_udf;
 use color::{lerp_color::make_lerp_color_udf, rgba::make_rgba_udf};
 use datafusion::logical_expr::ScalarUDF;
 use datafusion::prelude::SessionContext;
@@ -73,4 +76,6 @@ pub fn register_extension_udfs(ctx: &SessionContext) {
 
     ctx.register_udf(make_rgba_udf());
     ctx.register_udf(make_lerp_color_udf());
+
+    ctx.register_udf(make_bin_center_udf());
 }